moved GUI related actions and reducers to queueGUI instead of queue
